+++ doc_uploader/doc_processor/adapters/obsidian/__init__.py
