+++ doc_uploader/doc_processor/adapters/obsidian/processors.py
