+++ doc_uploader/doc_processor/adapters/template/__init__.py
